refactor: move LineClamp into SharedComponents
