+++ src/components/Posts/PostList/UpcomingWebinarListItem/UpcomingWebinarListItem.tsx
@@ -1,12 +1,11 @@
 import { FC } from "react";
 import { useHover } from "react-aria";
 
-import LineClamp from "../../../LineClamp";
-import OakLink from "../../../OakLink";
-import formatDate from "../../../../utils/formatDate";
-import { PostListItemProps } from "../PostListItem";
-import useClickableCard from "../../../../hooks/useClickableCard";
-
+import LineClamp from "@/components/SharedComponents/LineClamp";
+import OakLink from "@/components/OakLink";
+import formatDate from "@/utils/formatDate";
+import { PostListItemProps } from "@/components/Posts/PostList/PostListItem";
+import useClickableCard from "@/hooks/useClickableCard";
 import { Heading, P } from "@/components/SharedComponents/Typography";
 import ButtonAsLink from "@/components/SharedComponents/Button/ButtonAsLink";
 import Box from "@/components/SharedComponents/Box";
